+++ demo/demo/accounts/authentication.py
@@ -1,8 +1,0 @@
-from rest_framework.authentication import TokenAuthentication
-
-from .models import APIToken
-
-
-class APITokenAuthentication(TokenAuthentication):
-    keyword = 'Bearer'
-    model = APIToken