+++ tests/functional/test_email.py
@@ -2,13 +2,20 @@ from datetime import datetime
 
 from django.urls import reverse
 
+import pytest
+
 from demo.accounts.models import User, APIToken
 
 
-def test_signup(db, django_app, mailoutbox):
-    params = {"email": "zack@example.com", "password": "correct battery horse staple"}
-    django_app.post_json(reverse("auth:signup"), params=params, status=201)
+@pytest.mark.parametrize("email", [
+    "zack@example.com",
+    "zack@EXAMPLE.COM",
+])
+def test_signup(db, django_app, mailoutbox, email):
+    params = {"email": email, "password": "correct battery horse staple"}
+    resp = django_app.post_json(reverse("auth:signup"), params=params, status=201)
 
+    assert resp.json == {"email": "zack@example.com"}
     user = User.objects.last()
     assert user.email == "zack@example.com"
     assert user.has_usable_password()
@@ -38,17 +45,45 @@ def test_signup_invalid_email(db, django_app):
     assert "email" in resp.json
 
 
-def test_signup_already_exists(db, django_app, user0):
-    params = {"email": "bob@example.com", "password": "pass123"}
+@pytest.mark.parametrize("email", [
+    "bob@example.com",
+    "bob@example.COM",
+])
+def test_signup_already_exists(db, django_app, user0, email):
+    params = {"email": email, "password": "pass123"}
     resp = django_app.post_json(reverse("auth:signup"), params=params, status=400)
 
     assert "email" in resp.json
 
 
-def test_login(db, django_app, user0):
+def test_signup_different_address(db, django_app, user0):
+    params = {"email": "ZACK@EXAMPLE.com", "password": "goodpass"}
+    resp = django_app.post_json(reverse("auth:signup"), params=params, status=201)
+
+    assert resp.json == {"email": "ZACK@example.com"}
+
+
+def test_confirm_email(db, django_app, user0, emailconfirmation0):
+    assert not user0.is_active
+    assert emailconfirmation0.confirmed is None
+
+    django_app.get(reverse("app-auth:email-confirmation",
+                           kwargs={"external_id": emailconfirmation0.external_id}))
+
+    user0.refresh_from_db()
+    emailconfirmation0.refresh_from_db()
+    assert user0.is_active
+    assert isinstance(emailconfirmation0.confirmed, datetime)
+
+
+@pytest.mark.parametrize("email", [
+    "bob@example.com",
+    "bob@example.COM",
+])
+def test_login(db, django_app, user0, email):
     assert user0.last_login is None
 
-    params = {"email": "bob@example.com", "password": "pass123"}
+    params = {"email": email, "password": "pass123"}
     resp = django_app.post_json(reverse("auth:login"), params=params, status=200)
 
     assert "token" in resp.json