+++ demo/demo/pages/urls.py
@@ -1,4 +1,4 @@
-from django.conf.urls import url
+from django.urls import path
 
 from . import views
 
@@ -7,6 +7,6 @@ app_name = 'pages'
 
 
 urlpatterns = [
-    url(r'^$', views.index, name='root'),
-    url(r'^welcome/(?P<token>[^/.]+)$', views.confirm_email, name='confirm-email'),
+    path('', views.index, name='root'),
+    path('welcome/<token>/', views.confirm_email, name='confirm-email'),
 ]