+++ rest_auth_toolkit/__init__.py
@@ -1,5 +1,5 @@
 """Simple + flexible signup and login for Django APIs"""
 
-__version__ = '0.11.dev'
+__version__ = '0.11'
 
 default_app_config = 'rest_auth_toolkit.app.RestAuthToolkitConfig'