+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog for Rest-Framework-Auth-Toolkit
 
-## v0.11 (unreleased)
+## v0.11
 
-Require Django 2.2 (LTS release) or 3.1
-and Python 3.8 or 3.9 #115
+Tests are run with combinations of Python 3.8 and 3.9
+and Django 2.2 and 3.1 (#115)
 
 A number of improvements make it easier to integrate in your
 project (#135):
@@ -26,6 +26,8 @@ has a `last_login` field, it will be automatically updated.
 
 `BaseUserEmail.natural_key` now returns a 1-element tuple
 with the email field value.
+
+Use Python 3.8 or 3.9 and Django 2.2 (LTS release) or 3.1.
 
 
 ## v0.10